api/requirements.txt
Django==4.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
flake8==6.0.0
isort==5.12.0
PyYAML==6.0
requests==2.28.2
ujson==5.7.0
redis==4.5.5
eventlet==0.33.3
pytest-django==4.5.2
model_bakery==1.11.0
celery==5.2.7
pytest-cov==4.0.0
